+++ Foot/src/solveur/SolutionInitiale.java
@@ -5,7 +5,7 @@ import java.util.Collections;
 
 public class SolutionInitiale {
 	private int nbGroupe;
-	private int[][] tabDistance;
+	private double[][] tabDistance;
 	private int[] clubs;
 	private int nbClub;
 	private int[] solution;
@@ -21,7 +21,7 @@ public class SolutionInitiale {
 	 * @param dist
 	 * @return l'indice du club le plus éloigné (en moyenne)
 	 */
-	public int clubPlusLoin(int[][] dist, ArrayList<Integer> clubsAffecte) {
+	public int clubPlusLoin(double[][] dist, ArrayList<Integer> clubsAffecte) {
 		int[] somLign = new int[nbClub];
 
 		for (int i = 0; i < nbClub; i++) {
@@ -66,10 +66,10 @@ public class SolutionInitiale {
 	 * @param n
 	 * @return les n clubs les plus proches d'un club
 	 */
-	public int[] nClubsPlusProche(int n, int club, int[][] distance) {
-		ArrayList<Integer> distVoisins = new ArrayList<>();
+	public int[] nClubsPlusProche(int n, int club, double[][] distance) {
+		ArrayList<Double> distVoisins = new ArrayList<>();
 		for (int i = 0; i < n; i++) {
-			distVoisins.add(Integer.MAX_VALUE);
+			distVoisins.add(Double.MAX_VALUE);
 		}
 
 		int[] list = new int[n];
@@ -127,16 +127,16 @@ public class SolutionInitiale {
 		}
 	}
 
-	public SolutionInitiale(int nbGroupe, int[][] tabDistance, int nbClub) {
+	public SolutionInitiale(int nbGroupe, double[][] tabDistance, int nbClub) {
 		this.nbGroupe = nbGroupe;
 		this.tabDistance = tabDistance;
 		this.clubs = new int[nbClub];
 		this.nbClub = nbClub;
 		this.solution = new int[nbClub];
 
-		ArrayList<ArrayList<Integer>> ListDistance = new ArrayList<>();
+		ArrayList<ArrayList<Double>> ListDistance = new ArrayList<>();
 		for (int i = 0; i < nbClub; i++) {
-			ArrayList<Integer> listTemp = new ArrayList<>();
+			ArrayList<Double> listTemp = new ArrayList<>();
 
 			for (int j = 0; j < nbClub; j++) {
 				listTemp.add(j, tabDistance[i][j]);
@@ -149,7 +149,7 @@ public class SolutionInitiale {
 		for (int i = 0; i < nbGroupe; i++) {
 			int nbVoisins = nbClub / nbGroupe - 1;
 
-			int[][] tabDist = new int[ListDistance.size()][ListDistance.size()];
+			double[][] tabDist = new double[ListDistance.size()][ListDistance.size()];
 			for (int k = 0; k < ListDistance.size(); k++) {
 				for (int j = 0; j < ListDistance.size(); j++) {
 					tabDist[k][j] = ListDistance.get(k).get(j);
@@ -169,14 +169,14 @@ public class SolutionInitiale {
 			// maj ListDist des clubs utilisés à -1
 			for (int k : voisins) {
 				for (int j = 0; j < nbClub; j++) {
-					ListDistance.get(k).set(j, -1);
-					ListDistance.get(j).set(k, -1);
+					ListDistance.get(k).set(j, -1.0);
+					ListDistance.get(j).set(k, -1.0);
 				}
 
 			}
 			for (int j = 0; j < nbClub; j++) {
-				ListDistance.get(clubPlusLoin).set(j, -1);
-				ListDistance.get(j).set(clubPlusLoin, -1);
+				ListDistance.get(clubPlusLoin).set(j, -1.0);
+				ListDistance.get(j).set(clubPlusLoin, -1.0);
 			}
 
 		}
